backend/requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
psycopg[binary]
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose==3.3.0
